Add a PostV2Registry method for push authorization
This commit adds a method nearly identical to PingV2Registry,
except it uses a POST request instead of a GET request. This is
so that registries that use a proxy that perform authentication on
POST requests but not GET requests (to allow for anonymous pulls)
are able to provide this functionality. Previously, using a GET request
meant there was no way to differentiate if the action being taken was
a push or pull and handle it differently, this allows that.